% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webhook_data_io.R
\name{zappier_data_ingest}
\alias{zappier_data_ingest}
\alias{hcazapier}
\alias{write_session_data}
\alias{verifyPartnerKey}
\title{Webhook Data Integration}
\usage{
hcazapier(..., apikey = NULL)

write_session_data(file_name, input_data, apikey)

verifyPartnerKey(apikey)
}
\arguments{
\item{...}{data sent in request}

\item{apikey}{partner issued api key}
}
\description{
Endpoint to receive data external system, triggered by webhook,
must provide valid partner issued api key
}
\section{Functions}{
\itemize{
\item \code{hcazapier}: TBD

\item \code{write_session_data}: TBD

\item \code{verifyPartnerKey}: TBD
}}

